requirements.txt
-f https://git.io/purplship
purplship.package
django
djangorestframework
drf-yasg
djangorestframework-camel-case
